BigIntegers, option to print sequence or specified term, and commas
I converted the use of longs in the calculations to BigIntegers, so the
program can calculate the Fibonacci sequence at much greater term
numbers & values. I created and incorporated a function to add commas
onto term numbers, e.g. 1000th term --> 1,000th term. The user can now
select to only print the value at a specified term in the sequence, or
all the values up to and including the specified term. I'm planning on
also allowing the user to choose the format of the output FIbonacci
values - as plain numbers, with commas, or in scientific notation, as
well as maybe also adding a timer so the user can see how much time a
program run takes. I might also change the method of getting the term
number from taking in a long to taking in a BigInteger, to allow the
program to work to much greater term numbers than it currently can.